+++ library/src/BasisTheory.ts
@@ -6,6 +6,9 @@ import { BasisTheoryTokens } from './tokens';
 import { BasisTheoryEncryptionAdapters } from './encryption/BasisTheoryEncryptionAdapters';
 import type { BasisTheoryElements, BasisTheoryElementsInit } from './elements';
 import { BasisTheoryLogs } from './logs';
+import { BasisTheoryReactorFormulas } from './reactor-formulas';
+import { BasisTheoryReactors } from './reactors';
+import { BasisTheoryPermissions } from './permissions';
 
 export const defaultInitOptions: Required<BasisTheoryInitOptions> = {
   environment: 'production',
@@ -21,6 +24,9 @@ export class BasisTheory {
   private _elements?: BasisTheoryElements;
   private _applications?: BasisTheoryApplications;
   private _logs?: BasisTheoryLogs;
+  private _reactorFormulas?: BasisTheoryReactorFormulas;
+  private _reactors?: BasisTheoryReactors;
+  private _permissions?: BasisTheoryPermissions;
 
   public async init(
     apiKey: string,
@@ -52,6 +58,18 @@ export class BasisTheory {
       this._logs = new BasisTheoryLogs({
         apiKey,
         baseURL: SERVICES.logs[this._initOptions.environment],
+      });
+      this._reactorFormulas = new BasisTheoryReactorFormulas({
+        apiKey,
+        baseURL: SERVICES.reactorFormulas[this._initOptions.environment],
+      });
+      this._reactors = new BasisTheoryReactors({
+        apiKey,
+        baseURL: SERVICES.reactors[this._initOptions.environment],
+      });
+      this._permissions = new BasisTheoryPermissions({
+        apiKey,
+        baseURL: SERVICES.permissions[this._initOptions.environment],
       });
 
       this._encryption = new BasisTheoryEncryptionAdapters();
@@ -106,5 +124,17 @@ export class BasisTheory {
 
   public get logs(): BasisTheoryLogs {
     return assertInit(this._logs);
+  }
+
+  public get reactorFormulas(): BasisTheoryReactorFormulas {
+    return assertInit(this._reactorFormulas);
+  }
+
+  public get reactors(): BasisTheoryReactors {
+    return assertInit(this._reactors);
+  }
+
+  public get permissions(): BasisTheoryPermissions {
+    return assertInit(this._permissions);
   }
 }