+++ library/src/types.ts
@@ -8,7 +8,15 @@ export type ServiceUrlMap = {
   [key in ServiceEnvironment]: string;
 };
 
-export type Services = 'tokens' | 'atomic' | 'applications' | 'logs';
+export type Services =
+  | 'tokens'
+  | 'atomic'
+  | 'applications'
+  | 'reactorFormulas'
+  | 'reactors'
+  | 'permissions'
+  | 'logs';
+
 export type ServicesMap = {
   [key in Services]: ServiceUrlMap;
 };