+++ library/test/environment.test.ts
@@ -43,7 +43,19 @@ describe('Environments', () => {
       ...baseConfig,
       baseURL: SERVICES.logs[environment],
     });
-    expect(create).toHaveBeenCalledTimes(4);
+    expect(create).toHaveBeenCalledWith({
+      ...baseConfig,
+      baseURL: SERVICES.reactorFormulas[environment],
+    });
+    expect(create).toHaveBeenCalledWith({
+      ...baseConfig,
+      baseURL: SERVICES.reactors[environment],
+    });
+    expect(create).toHaveBeenCalledWith({
+      ...baseConfig,
+      baseURL: SERVICES.permissions[environment],
+    });
+    expect(create).toHaveBeenCalledTimes(7);
   });
 
   it('should throw error if not properly initialized', () => {